Handle busy state on deletion

+++ modules/start-and-stop/function/src/delete-runner-helper.js
@@ -5,22 +5,32 @@ const utils = require('./utils')
 const chalk = require('chalk')
 
 module.exports.deleteRunner = deleteRunner
+module.exports.deleteRunnerFromGitHub = deleteRunnerFromGitHub
 module.exports.deleteRunnerVm = deleteRunnerVm
 
 async function deleteRunner (runnerName) {
   console.info(`start delete runner ${runnerName}...`)
-  await deleteRunnerFromGitHub(runnerName)
-  await deleteRunnerVm(runnerName)
-  console.info(chalk.green(`runner ${runnerName} is fully deleted`))
+  const deleteFromGitHubSucceed = await deleteRunnerFromGitHub(runnerName)
+  if (deleteFromGitHubSucceed) {
+    await deleteRunnerVm(runnerName)
+    console.info(chalk.green(`runner ${runnerName} is fully deleted`))
+  } else {
+    console.info(chalk.yellow(`runner ${runnerName} was not deleted`))
+  }
 }
 
 async function deleteRunnerFromGitHub (runnerName) {
   const githubStatus = await githubHelper.getGitHubRunnerByName(runnerName)
-  if (githubStatus !== null) {
+  if (githubStatus === null) {
+    throw Error(`runner ${runnerName} is unknown from github`)
+  }
+  if (githubStatus.busy === false) {
     await githubHelper.deleteGitHubRunner(githubStatus.id)
     console.info(`runner ${runnerName} deleted from GitHub`)
+    return true
   } else {
-    console.warn(chalk.yellow(`runner GitHub status for ${runnerName} is unknown`))
+    console.warn(chalk.yellow(`runner ${runnerName} is busy, can not delete it form github`))
+    return false
   }
 }
